Begin modulatting WooMinecraft::check() method
This method is WAY too huge and needs to be updated.  To date I'm breaking
out the connection portion of the method into its own with more try..catch
calls.  Also I've broken out the player list generator, so it will be a
lot easier to modify at a later date.